refactor: 로또 티켓 관련 로직을 ticketView로 분리
