+++ src/step2-index.js
@@ -2,10 +2,11 @@
  * step 2의 시작점이 되는 파일입니다.
  * 노드 환경에서 사용하는 readline 등을 불러올 경우 정상적으로 빌드할 수 없습니다.
  */
-import { SETTINGS } from "./constants/index.js";
 import LottoController from "./controller/LottoController.js";
 import { initPurchaseUI, resetPurchaseUI } from "./view/purchaseView.js";
+import { resetLottoTicketsUI } from "./view/ticketView.js";
 import { showModal, updateModalContent, restartModal } from "./view/modalView.js";
+import { SETTINGS } from "./constants/index.js";
 
 const lottoController = new LottoController();
 
@@ -71,6 +72,7 @@ winningResultButton.addEventListener("click", () => {
 
 const resetModalUI = () => {
   resetPurchaseUI();
+  resetLottoTicketsUI();
   winningNumberInputs.forEach((input) => input.value = "");
   bonusInput.value = "";
 };
